# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restaurantes API",
        "contact": {
            "url": "",
            "email": ""
        },
        "license": {
            "name": "Deber de Programacion Web"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/platos": {
            "get": {
                "tags": [
                    "platos"
                ],
                "summary": "Obtener lista de platos",
                "operationId": "getPlatosList",
                "responses": {
                    "200": {
                        "description": "Lista de platos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Platos"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "platos"
                ],
                "summary": "Crear un nuevo plato",
                "operationId": "createPlato",
                "requestBody": {
                    "description": "Datos del plato",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Platos"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Un plato",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Platos"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/platos/{id}": {
            "get": {
                "tags": [
                    "platos"
                ],
                "summary": "Obtener plato por id",
                "operationId": "getPlatoById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del plato",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Un plato"
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "platos"
                ],
                "summary": "Actualizar un plato",
                "operationId": "updatePlato",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del plato",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos del plato",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Platos"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Un plato",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Platos"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "platos"
                ],
                "summary": "Eliminar un plato",
                "operationId": "deletePlato",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del plato",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Un plato",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Platos"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/restaurantes": {
            "get": {
                "tags": [
                    "restaurantes"
                ],
                "summary": "Lista de restaurantes",
                "operationId": "listRestaurantes",
                "responses": {
                    "200": {
                        "description": "Un array de restaurantes",
                        "headers": {
                            "x-next": {
                                "description": "A link to the next page of responses",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Restaurantes"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "restaurantes"
                ],
                "summary": "Crear un restaurante",
                "operationId": "createRestaurante",
                "requestBody": {
                    "description": "Datos del restaurante",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Restaurantes"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Un restaurante"
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/restaurantes/{id}": {
            "get": {
                "tags": [
                    "restaurantes"
                ],
                "summary": "Obtener restaurante por id",
                "operationId": "getRestauranteById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del restaurante",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Un restaurante"
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "restaurantes"
                ],
                "summary": "Actualizar un restaurante",
                "operationId": "updateRestaurante",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del restaurante",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos del restaurante",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Restaurantes"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Un restaurante"
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "restaurantes"
                ],
                "summary": "Eliminar un restaurante",
                "operationId": "deleteRestaurante",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del restaurante",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Un restaurante"
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Error": {
                "title": "Error",
                "description": "Error",
                "properties": {
                    "code": {
                        "description": "Codigo de error",
                        "type": "integer",
                        "format": "int32",
                        "example": 400
                    },
                    "message": {
                        "description": "Mensaje de error",
                        "type": "string",
                        "example": "Error de validacion"
                    },
                    "errors": {
                        "description": "Errores",
                        "type": "array",
                        "items": {
                            "description": "Error",
                            "type": "string",
                            "example": "El campo nombre es requerido"
                        }
                    }
                },
                "type": "object"
            },
...
            "Platos": {
                "title": "Platos",
                "description": "Platos",
                "properties": {
                    "nombre": {
                        "description": "Nombre",
                        "type": "string",
                        "example": "Plato 1"
                    },
                    "precio": {
                        "description": "Precio",
                        "type": "string",
                        "example": "10.00"
                    },
                    "descripcion": {
                        "description": "Descripcion",
                        "type": "string",
                        "example": "Descripcion del plato 1"
                    },
                    "id_restaurante": {
                        "description": "ID del restaurante",
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
            "Restaurantes": {
                "title": "Restaurantes",
                "description": "Restaurantes",
                "properties": {
                    "ruc": {
                        "description": "RUC",
                        "type": "string",
                        "example": "12345678901"
                    },
                    "nombre": {
                        "description": "Nombre",
                        "type": "string",
                        "example": "Restaurante 1"
                    },
                    "direccion": {
                        "description": "Direccion",
                        "type": "string",
                        "example": "Av. Los Olivos 123"
                    },
                    "telefono": {
                        "description": "Telefono",
                        "type": "string",
                        "example": "987654321"
                    },
                    "email": {
                        "description": "Email",
                        "type": "string",
                        "example": "restaurante@email.com"
                    },
                    "categoria": {
                        "description": "Categoria",
                        "type": "string",
                        "example": "Comida rapida"
                    }
                },
                "type": "object"
            },
...
        }
    }
}